+++ example/src/gh-issues/Issue1098NestedPagerRepro.tsx
